Revert "fix uint int comparisons in rendermodel"
This reverts commit 9df6e44f3ba1f06cb15b24b4ccd555ae1ac40b44.

+++ src/engine/render/rendermodel.cpp
@@ -682,7 +682,7 @@ void resetmodelbatches()
 void addbatchedmodel(model *m, batchedmodel &bm, int idx)
 {
     modelbatch *b = nullptr;
-    if(static_cast<int>(batches.size()) > m->batch)
+    if(batches.size() > m->batch)
     {
         b = &batches[m->batch];
         if(b->m == m && (b->flags & Model_Mapmodel) == (bm.flags & Model_Mapmodel))
@@ -1100,7 +1100,7 @@ void startmodelquery(occludequery *query)
 
 void endmodelquery()
 {
-    if(static_cast<int>(batchedmodels.size()) == modelquerymodels)
+    if(batchedmodels.size() == modelquerymodels)
     {
         modelquery->fragments = 0;
         modelquery = nullptr;
